layers: Move PipelineLayout CoreCheck to Descriptor file
